\name{analysis-class}
\Rdversion{1.1}
\docType{class}
\alias{analysis-class}

\title{Class \code{"analysis"}}
\description{...}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{...}
\seealso{...}
\examples{
showClass("analysis")
}
\keyword{classes}
