\name{asFile-methods}
\docType{methods}
\alias{asFile-methods}
\alias{asFile,ANY-method}
\alias{asFile,Matrix-method}
\alias{asFile,character-method}
\alias{asFile,list-method}
\alias{asFile,matrix-method}

\title{methods for function \code{asFile} in package \pkg{matR}}
\description{...}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{...}
\item{\code{signature(x = "Matrix")}}{...}
\item{\code{signature(x = "character")}}{...}
\item{\code{signature(x = "collection")}}{...}
\item{\code{signature(x = "list")}}{...}
\item{\code{signature(x = "matrix")}}{...}
}}
\seealso{\code{\link{asFile}}}
\keyword{methods}
