\name{asFile}
\alias{asFile}
\title{Export Objects from matR}
\description{
	Simply exports various kinds of object to a file
}
\usage{
asFile (x, file, ...)
}
\arguments{
	\item{x}{...}
	\item{file}{...}
	\item{...}{...}
}
\details{
...
}
\value{
...
}
\author{Daniel T. Braithwaite}
\seealso{\code{\link{asFile-methods}}, \code{\link{msession}}, \code{\link{write.table}}, \code{\link{save}}}
\examples{
#
}
