\name{boxplot}
\alias{boxplot}

\title{Boxplot of Collection Annotations}
\description{...}
\usage{boxplot(x, ...)}
\arguments{
  \item{x}{collection}
  \item{\dots}{further arguments as indicated below}
}
%\details{...}
%\value{...}
\author{Kevin P. Keegan, Daniel T. Braithwaite}
\seealso{\code{\link{boxplot-methods}}}
\examples{##}
