\name{buildDemoSets}
\alias{buildDemoSets}

\title{Build example datasets}

\description{Build example datasets.}

\usage{
buildDemoSets(file = "sample-sets.rda")
}

\arguments{
  \item{file}{filename for output ("character")}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
