% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_operations.R
\name{capture_images}
\alias{capture_images}
\title{Capture Images from a Webcam}
\usage{
capture_images(path, n = 50, device = NULL)
}
\arguments{
\item{path}{Where the images will be stored}

\item{n}{The number of images to capture}
}
\value{
TRUE if successful
}
\description{
Capture Images from a Webcam
}
