\name{collapse}
\alias{collapse}

\title{Collapse character vectors to a single string}

\description{Collapse one or more character vectors to a single string, with an optional separator.}

\usage{
collapse(x, ..., sep=" ")
}

\arguments{
  \item{x}{a \code{character} vector}
  \item{\dots}{more \code{character} vectors}
  \item{sep}{a separator (\code{character})}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
