\name{collection-methods}
\docType{methods}
\alias{collection-methods}
\alias{collection,ANY-method}
\alias{collection,selection-method}

\title{methods for function \code{collection} in package \pkg{matR}}
\description{...}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{...}
\item{\code{signature(x = "selection")}}{...}
}}
\seealso{\code{\link{Collections}}}
\keyword{methods}
