\name{demo.step}
\alias{demo.step}

\title{Step through a demo line-by-line}

\description{Step through a demo line-by-line.}

\usage{
demo.step(s)
}

\arguments{
  \item{s}{path to source file ("character")}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
