\name{demo2}
\alias{demo2}
\alias{Demos}
\alias{demo}
\alias{demo1}
\title{
Run Demos of matR Functionality
}
\description{
These functions step through demonstrations of \code{matR} functionality.
}
\usage{
demo1()
demo2()
}
\details{
  The demos presented by these functions can also be run through the standard \code{R} demo mechanism, \code{demo()},
  as \code{demo(first)} and \code{demo(second)}.  These functions slightly improve on that mechanism, however, by running
  the demos step-by-step instead of all at once.
}
\value{
}
\references{
}
\author{
Daniel Braithwaite
}
\note{
}
\seealso{
}
\examples{
}
