\name{Demos}
\alias{Demos}
\alias{demos}
\alias{Demo}
\alias{demo}
\alias{demo1}
\alias{demo2}
\title{Run Demos of matR Functionality}
\description{
These functions step through demonstrations of \code{matR} functionality, including downloads of metagenome collections, metadata, analyses, and visualizations.
}
\usage{
demo1()
demo2()
}
\details{
These functions improve on \code{demo()}, the standard demo mechanism in \code{R}, by proceeding stepwise (prompting at each line).  
(However, \code{demo(first)} and \code{demo(second)} can be used to run the same demos using the standard mechanism.)
}
\author{
Daniel T. Braithwaite
}
