\name{ExampleData}
\alias{ExampleData}
\alias{Example Data}
\alias{exampleMetadata}
\alias{exampleMatrix}
\alias{exampleIDs}
\docType{data}
\title{
Example Kbase Datasets
}
\description{
A metagenome metadata header, a taxon abundance matrix, and a vector of the associated metagenome IDs.
}
\usage{
data(exampleMetadata)
data(exampleMatrix)
data(exampleIDs)
}
\source{
\url{http://metagenomics.anl.gov}
}
\examples{
data(exampleMetadata)
}
\keyword{datasets}
