% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{fit_knn}
\alias{fit_knn}
\title{Create a KNN Instance}
\usage{
fit_knn(X, y, k = 6)
}
\arguments{
\item{X}{Training matrix}

\item{y}{Training classes}

\item{k}{Number of neighbors to consider}
}
\value{
A KNN instance
}
\description{
Set up a KNN instance which bundles the training data
}
