% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{fit_network}
\alias{fit_network}
\title{Fit a Neural Network Classifier}
\usage{
fit_network(X, Y, neurons = 3L, epoch = 1000L, alpha = 0.01, seed = 123)
}
\arguments{
\item{X}{Matrix of training data}

\item{Y}{Matrix of training labels}

\item{neurons}{Number of hidden layer neurons}

\item{epoch}{Number of learning epochs}

\item{alpha}{Learning rate}
}
\value{
A list as class visionary neural network
}
\description{
Fits a neural network using C++ backend.
}
