% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{fit_tree}
\alias{fit_tree}
\title{Fit a Binary Decision Tree Classifier}
\usage{
fit_tree(X, y, min_split = 100)
}
\arguments{
\item{X}{Matrix of training data}

\item{y}{Matrix of training labels}

\item{min_split}{Min number of samples to split}
}
\value{
A list as class deepspace decision tree
}
\description{
Fits a decision tree using C++ backend.
}
\examples{

data <- ggplot2::diamonds

y <- 1 * matrix(data$price > 2400)
X <- as.matrix(dplyr::select(data, carat, table))

tree <-
  deepspace::fit_tree(
    X = X,
    y = y,
    min_split = 100
  )

predict(tree)
}
