\name{getMetColumns}
\alias{getMetColumns}

\title{Get single field from column metadata}

\description{Get single field from column metadata.}

\usage{
getMetColumns(name, xx)
}

\arguments{
  \item{name}{string (\code{character})}
  \item{xx}{object (\code{biom})}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
