\name{hazPackages}
\alias{hazPackages}

\title{Check for presence of recommnded packages}

\description{Check for presence of recommnded packages.}

\usage{
hazPackages()
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
