\name{KbaseKit-package}
\alias{KbaseKit-package}
\alias{KbaseKit}
\alias{Kbase}
\alias{kbase}
\alias{mgrast}
\alias{MGRAST}
\alias{MG-RAST}
\docType{package}
\title{
Basic Toolkit for the Kbase MG-RAST Server
}
\description{
This package provides facilities to retrieve inputs and outputs of the Kbase MG-RAST analysis pipeline.  
Access to private data is authenticated.
Data is encapsulated with metadata in a lightweight object.
Import and export from BIOM format is enabled.
A suite of statistical tools is provided.
}
\details{
\tabular{ll}{
Package: \tab KbaseKit\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2012-02-27\cr
Depends: \tab RJSONIO, matlab, ecodist, Cairo, gplots\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
}
\author{
}
\references{
\code{\href{http://metagenomics.anl.gov}{http://metagenomics.anl.gov}}
}
\keyword{ package }
\keyword{ Kbase }
\keyword{ kbase }
\keyword{ KbaseKit }
\keyword{mgrast}
\keyword{MGRAST}
\keyword{MG-RAST}
\seealso{
\code{\link{Data Retrieval from MG-RAST}}\cr
\code{\link{Authorization}}\cr
\code{\link{Querying MG-RAST}}\cr
\code{\link{Import and Export}}\cr
\code{\link{Analysis Tools}}\cr
\code{\link{Direct API Calls}}\cr
\cr
%\code{\link{kbGet}},\cr
%\code{\link{kbListAll}},\cr
%\code{\link{kbSearchMetagenomes}},\cr
%\code{\link{kbAnnotationLookup}},\cr
\cr
%\code{\link{mgrastPreprocessing}},\cr
%\code{\link{mgrastSuggestTest}},\cr
%\code{\link{mgrastDoStats}},\cr
%\code{\link{mgrastSampleMatrix}},\cr
%\code{\link{mgrastDendrograms}},\cr
%\code{\link{mgrastPlotPCA}},\cr
%\code{\link{mgrastPlotPCO}}\cr
}
\examples{
}
