% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_operations.R
\name{live_stream}
\alias{live_stream}
\title{Initialize a Live Video Stream}
\usage{
live_stream(model, device = NULL)
}
\arguments{
\item{model}{A neural network model}
}
\value{
Nothing
}
\description{
Initialize a Live Video Stream
}
