\name{metadata-class}
\Rdversion{1.1}
\docType{class}
\alias{metadata-class}

\title{Class \code{"metadata"}}
\description{...}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}
\section{Methods}{
No methods defined with class "metadata" in the signature.
}
\seealso{
	\code{\link{metadata}}
}
\examples{
showClass("metadata")
}
\keyword{classes}
