% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{new_matrix}
\alias{new_matrix}
\title{Create a Matrix}
\usage{
new_matrix(...)
}
\arguments{
\item{...}{Vectors as rows of the matrix}
}
\value{
A matrix
}
\description{
A NumPy like interface for building matrices
}
\examples{

matr::new_matrix(
  c(1, 2, 3),
  c(3, 2, 1),
  c(1, 1, 2)
)

}
