\name{normalize}
\alias{normalize}
\title{Centered and Scaled Logarithm of Abundance Data}
\description{Centered and Scaled Logarithm of Abundance Data}
\usage{
normalize(x, method = c("standard"), ...)
}
\arguments{
  \item{x}{...}
  \item{method}{...}
  \item{...}{...}
}
\details{...}
\value{...}
\author{Kevin P. Keegan, Daniel T. Braithwaite}
\examples{
#
}

