\name{parMap}
\alias{parMap}

\title{Determine par values according to a provided metadata mapping}

\description{Determine par values according to a provided metadata mapping.}

\usage{
parMap(object, name.map, value.map)
}

\arguments{
  \item{object}{object with metadata (\code{biom})}
  \item{name.map}{mapping of par names to metadata names (\code{character})}
  \item{value.map}{list possibly including mappings of metadata values to par values (\code{list})}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
