\name{parMapper}
\alias{parMapper}

\title{Determine par values according to a provided metadata mapping}

\description{Determine par values according to a provided metadata mapping.}

\usage{
parMapper(object, name.map, value.map)
}

\arguments{
  \item{object}{object with metadata ("biom")}
  \item{name.map}{mapping of par names to metadata names ("character")}
  \item{value.map}{list possibly including mappings of metadata values to par values ("list")}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
