\name{pco-methods}
\docType{methods}
\alias{pco-methods}
\alias{pco,ANY-method}
\alias{pco,collection-method}
\title{Methods for Function \code{pco} in Package \pkg{matR}}
\description{...}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{...}
\item{\code{signature(x = "collection")}}{...}
}}
\seealso{\code{\link{pco}}}
\keyword{methods}
