\name{readSet}
\alias{readSet}

\title{Read sample and project IDs from a file}

\description{Read sample and project IDs from a file.}

\usage{
readSet(file)
}

\arguments{
  \item{file}{a filename ("character")}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
