\name{remove.singletons}
\alias{remove.singletons}
\title{Clean Up Matrix of Abundance Data}
\description{Replaces NA and single counts with zero, and removes rows of zeros}
\usage{remove.singletons(x, lim.entry = 1, lim.row = 1, ...)}
\arguments{
  \item{x}{...}
  \item{lim.entry}{...}
  \item{lim.row}{...}
  \item{...}{...}
}
\details{...}
\value{...}
\author{Kevin P. Keegan, Daniel T. Braithwaite}
\examples{
#
}
