\name{render}
\alias{render}
\alias{Visualization}
\alias{visualization}
\title{Produce Graphical Representation}
\description{
  \code{render} is a generic function to put analyses into graphical form.
  Current methods reproduce visualizations also implemented on the \code{MG-RAST} analysis webpages.
}
\usage{
render (x, ...)
}
\arguments{
  \item{x}{...}
  \item{...}{...}
}
\details{

Details of graphical parameters.
}
\value{
}
\references{
}
\author{Daniel T. Braithwaite}
\seealso{\code{\link{render-methods}}, \code{\link{collection}}, \code{\link{pco}}, \code{\link{heatmap}}}
\examples{
}
