\name{rownames-methods}
\docType{methods}
\alias{rownames-methods}
\alias{rownames,ANY-method}
\title{Methods for Function \code{rownames} in Package \pkg{matR}}
\description{...}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{...}
\item{\code{signature(x = "collection")}}{...}
}}
\keyword{methods}
