\name{scrapeSet}
\alias{scrapeSet}

\title{Scrape resource types from a list of IDs}

\description{Scrape resource types from a list of IDs.}

\usage{
scrapeSet(ids)
}

\arguments{
  \item{ids}{one or more resource ids ("character")}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
