\name{scrubSet}
\alias{scrubSet}

\title{Scrub a list of IDs to standard format}

\description{Scrub a list of IDs to standard format.}

\usage{
scrubSet(IDs, resources = "metagenome")
}

\arguments{
  \item{IDs}{one or more character strings ("character")}
  \item{resources}{corresponding resource designation(s) ("character")}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
