\name{selection-methods}
\docType{methods}
\alias{selection-methods}
\alias{selection,character-method}
\alias{selection,collection-method}
\alias{selection,numeric-method}
\alias{selection}
\title{Methods for Function \code{selection} in Package \pkg{matR}}
\description{...}
\section{Methods}{
\describe{
\item{\code{signature(x = "character")}}{...}
\item{\code{signature(x = "collection")}}{...}
\item{\code{signature(x = "numeric")}}{...}
\item{\code{signature(x = "selection")}}{...}
}}
\seealso{\code{\link{selection-class}}}
\keyword{methods}
