\name{sigtest-methods}
\docType{methods}
\alias{sigtest-methods}
\alias{sigtest,collection-method}
\alias{sigtest,matrix-method}
\title{Methods for Function \code{sigtest} in Package \pkg{matR}}
\description{...}
\section{Methods}{
\describe{
\item{\code{signature(x = "collection")}}{...}
\item{\code{signature(x = "matrix")}}{...}
}}
\seealso{\code{\link{sigtest}}}
\keyword{methods}
