\name{step.through}
\alias{step.through}

\title{Step through a demo line-by-line}

\description{Step through a demo line-by-line.}

\usage{
step.through(s)
}

\arguments{
  \item{s}{path to source file ("character")}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
