\name{subMetColumns}
\alias{subMetColumns}

\title{Substitute for reference to column metadata}

\description{Substitute for reference to column metadata.}

\usage{
subMetColumns(name, xx)
}

\arguments{
  \item{name}{a metadata reference ("character")}
  \item{xx}{an object ("biom")}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
