\name{subMetRows}
\alias{subMetRows}

\title{Substitute for reference to row metadata}

\description{Substitute for reference to row metadata.}

\usage{
subMetRows(name, xx)
}

\arguments{
  \item{name}{a metadata reference ("character")}
  \item{xx}{an object ("biom")}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
