% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{train_test}
\alias{train_test}
\title{Train Test Split}
\usage{
train_test(X, Y, train_prop = 0.6, seed = 123)
}
\arguments{
\item{X}{Matrix of training data}

\item{Y}{Matrix of training labels}

\item{train_prop}{Proportion of data for training}

\item{seed}{Random seed value}
}
\value{
A list of matrices
}
\description{
Splits X and Y data matrices into train and test
}
\examples{
data <- matr:::mnist

data_split <-
  matr::train_test(
    X = data$X,
    Y = data$Y
  )

}
