\name{analysis-class}
\alias{analysis-class}
\alias{Analyses}
\alias{analyses}
\docType{class}

\title{Class \code{"analysis"}}
\description{
A virtual class defined to encapsulate analyses defined by the \code{matR} pacakge.  [In principle all analyses can be saved with \code{asFile} and visualized with \code{render}]
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "analysis" in the signature.
}
\author{Daniel T. Braithwaite}
\seealso{
\code{\link{pco}},\code{\link{heatmap}},\code{\link{render}}}
\keyword{classes}
