\name{analysis-class}
\Rdversion{1.1}
\docType{class}
\alias{analysis-class}

\title{class \code{"analysis"}}
\description{Virtual container for analysis products and visualization parameters}
\seealso{\code{\link{heatmap-class}},\code{\link{pco-class}},\code{\link{sigtest-class}}}
\keyword{classes}
