\name{buildDemoSets}
\alias{buildDemoSets}

\title{
Build objects containing metagenome annotations as BIOM data
}

\description{
Build (for packaging) or rebuild certain objects of class \code{biom} for demonstration purposes.
}

\usage{
buildDemoSets(n=TRUE, file="demoObjects.rda")
}

\arguments{
  \item{n}{which sets (\code{integer})}
  \item{file}{filename for output (\code{character})}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\seealso{
\code{\link{demoObjects}},
\code{\link{demoSets}}
}
\keyword{internal}
