\name{demo2}
\alias{Demos}
\alias{demo2}
\title{Step Through Demos Line by Line}
\description{Improves on standard function demo() by stepping line by line}
\usage{
demo2(s)
}
\arguments{
  \item{s}{...}
}
\details{See \code{demo (package="matR")} for a list of provided demos.}
\value{...}
\author{Daniel T. Braithwaite}
\seealso{...}
\examples{
\dontrun{
# step through metadata demo
demo2 ("metadata")

# step through second analysis demo
demo2 ("another")
}
}
