\name{demo2}
\alias{Demos}
\alias{demo2}
\title{Step Through Tutorial Demos}
\description{Improves on builtin function \code{demo()} by stepping line by line}
\usage{
demo2(s)
}
\arguments{
  \item{s}{name of demo to run (character)}
}
\details{See \code{demo(package="matR")} for a list of packaged demos.}
%\value{...}
\author{Daniel T. Braithwaite}
%\seealso{...}
\examples{
\dontrun{
# step through metadata demo
demo2("metadata")

# step through second analysis demo
demo2("another")
}
}
