\name{dependencies}
\alias{dependencies}

\title{Install recommended packages}

\description{Install recommended packages.}

\usage{
dependencies(prompt=TRUE)
}

\arguments{
  \item{prompt}{interactively? ("logical")}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}

\examples{
\dontrun{
dependencies()
}
}
\keyword{internal}