\name{dependencies}
\alias{dependencies}

\title{Install recommended packages}

\description{Install recommended packages.}

\usage{
dependencies(prompt=TRUE)
}

\arguments{
  \item{prompt}{interactively? (\code{logical})}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}