\name{ExampleData}
\alias{ExampleData}
\alias{Example Data}
\alias{exampleMetadata}
\alias{exampleMatrix}
\alias{exampleIDs}
\alias{samp}
\alias{guts}
\alias{waters}
\docType{data}
\title{
Example Data 
}
\description{
Various data objects predefined in \code{matR} for toying around.
}
\usage{
data(exampleMetadata)
data(exampleMatrix)
data(exampleIDs)
}
\source{
\url{http://metagenomics.anl.gov}
}
\examples{
\dontrun{
data(exampleMetadata)
}
}
\keyword{datasets}
