% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{fit_network}
\alias{fit_network}
\title{Fit a Neural Network Classifier}
\usage{
fit_network(X, Y, neurons = 3L, epoch = 250L, learn_rate = 1e-05, seed = 123)
}
\arguments{
\item{X}{Matrix of training data}

\item{Y}{Matrix of training labels}

\item{neurons}{Number of hidden layer neurons}

\item{epoch}{Number of learning epochs}

\item{learn_rate}{Learning rate}

\item{seed}{Random seed value}
}
\value{
A list as class matr neural network
}
\description{
Fits a neural network using C++ backend.
}
\examples{
data <- matr:::mnist

network <-
  matr::fit_network(
    X = data$X,
    Y = data$Y,
    neurons = 5,
    epoch = 1000,
    learn_rate = 0.0001
  )

predict(network)

}
