% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{fit_network}
\alias{fit_network}
\title{Fit a Neural Network}
\usage{
fit_network(X, Y, neurons = 5L, epoch = 10000L, alpha = 0.001)
}
\arguments{
\item{X}{Matrix of training data}

\item{Y}{Matrix of training labels}

\item{neurons}{Hidden layer neurons}

\item{epoch}{Number of epochs}

\item{alpha}{Learning rate}
}
\value{
Neural network
}
\description{
\eqn{x = 2}
\deqn{x = 3}
}
