% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{fit_network}
\alias{fit_network}
\title{Fit a Neural Network Classifier}
\usage{
fit_network(
  X,
  Y,
  neurons = 3L,
  epoch = 100L,
  learn_rate = 0.01,
  stop_tol = -Inf,
  seed = 123
)
}
\arguments{
\item{X}{Matrix of training data}

\item{Y}{Matrix of training labels}

\item{neurons}{Number of hidden layer neurons}

\item{epoch}{Number of learning epochs}

\item{learn_rate}{Learning rate}

\item{stop_tol}{The convergence tolerance limit}

\item{seed}{Random seed value}
}
\value{
A list as class deepspace neural network
}
\description{
Fits a neural network using C++ backend.
}
