% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_operations.R
\name{live_stream}
\alias{live_stream}
\title{Initialize a Live Video Stream}
\usage{
live_stream(model, fun = NULL, device = NULL)
}
\arguments{
\item{model}{A neural network model}

\item{device}{Path to linux image device}
}
\description{
Initialize a Live Video Stream
}
