% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_operations.R
\name{load_images}
\alias{load_images}
\title{Load Images from a Directory as Data}
\usage{
load_images(
  path,
  label = 1,
  x_range = seq(140, 499, 10),
  y_range = seq(1, 360, 10),
  pattern = ".jpg$"
)
}
\arguments{
\item{path}{A path to an image or directory of images}

\item{label}{A label (true label) for training}

\item{x_range}{Subset of x axis pixels for training}

\item{y_range}{Subset of y axis pixels for training}

\item{pattern}{A regex pattern for file matching}
}
\value{
A tibble
}
\description{
Load Images from a Directory as Data
}
