\name{matR-package}
\alias{matR-package}
\docType{package}
\title{Metagenomics Analysis Tools for R}
\description{
This package provides access to the MG-RAST metagenome annotation
pipeline and related analysis tools.
}
\details{
matR is an analysis client for MG-RAST built as
an extension package to the popular R
programming language for statistics.  This
package is in development.  Facilities are
implemented to download annotation data from the
MG-RAST metagenome analysis server.  Easy
acccess to lightly structured metadata,
authentication for access to private data, and
customized versions of relevant statistical and
visualization tools are provided.
}
\author{Daniel T. Braithwaite, Kevin P. Keegan}
\references{
\url{http://metagenomics.anl.gov}\cr
\url{http://www.kbase.us}\cr
}
\seealso{
\code{\link{Authentication}}\cr
\code{\link{Collections}}\cr
\code{\link{Demos}}\cr
\code{\link{Dependencies}}\cr
}
\keyword{package}
