\name{normalize}
\alias{normalize}
\title{Log-scale and Center Matrix}
\description{Centers and scales logarithm of entries of a numeric matrix}
\usage{
normalize(x, method = c("standard"), ...)
}
\arguments{
  \item{x}{matrix}
  \item{method}{currently unused}
  \item{...}{currently unused}
}
%\details{...}
%\value{...}
\author{Kevin P. Keegan, Daniel T. Braithwaite}
\examples{
#
}

