\name{parAuto}
\alias{parAuto}

\title{Provide automatic par values}

\description{Provide automatic par values.}

\usage{
parAuto(par.name, n)
}

\arguments{
  \item{par.name}{name of graphical parameter (\code{character})}
  \item{n}{how many values are needed (\code{integer})}
}

\details{Internal and undocumented at present.}
\value{Internal and undocumented at present. }
\author{Daniel T. Braithwaite}
\keyword{internal}
