\name{remove.singletons}
\alias{remove.singletons}
\title{Clean Up Matrix of Abundance Data}
\description{Replaces NA and single counts with zero, and removes rows of zeros}
\usage{remove.singletons(m, lim = 1)}
\arguments{
  \item{m}{...}
  \item{lim}{...}
}
\details{...}
\value{...}
\author{Kevin P. Keegan, Daniel T. Braithwaite}
\examples{
#
}
